+++ toolbox/fdc3-conformance/src/test/advanced/fdc3.findIntent.ts
@@ -1,17 +1,17 @@
 import { AppIntent, IntentMetadata, ResolveError, DesktopAgent } from '@finos/fdc3';
 import { assert, expect } from 'chai';
-import { APIDocumentation2_0 } from '../support/apiDocuments-2.0';
-import { ContextType, IntentApp, Intent } from '../support/intent-support-2.0';
+import { APIDocumentation } from '../support/apiDocuments';
+import { ContextType, IntentApp, Intent } from '../support/intent-support';
 
 declare let fdc3: DesktopAgent;
-const findIntentDocs = '\r\nDocumentation: ' + APIDocumentation2_0.findIntent + '\r\nCause';
+const findIntentDocs = '\r\nDocumentation: ' + APIDocumentation.findIntent + '\r\nCause';
 
 /**
  * Details on the mock apps used in these tests can be found in /mock/README.md
  */
 export default async () =>
   describe('fdc3.findIntent', () => {
-    it("(2.0-FindIntentAppD) Should find intent 'aTestingIntent' belonging only to app intent-a", async () => {
+    it("(FindIntentAppD) Should find intent 'aTestingIntent' belonging only to app intent-a", async () => {
       const appIntent = await fdc3.findIntent(Intent.aTestingIntent);
       validateAppIntent(
         appIntent,
@@ -21,7 +21,7 @@ export default async () =>
       );
     });
 
-    it('(2.0-FindNonExistentIntentAppD) Should throw NoAppsFound error when intent does not exist', async () => {
+    it('(FindNonExistentIntentAppD) Should throw NoAppsFound error when intent does not exist', async () => {
       try {
         await fdc3.findIntent('nonExistentIntent');
         assert.fail('No error was thrown', findIntentDocs);
@@ -30,7 +30,7 @@ export default async () =>
       }
     });
 
-    it("(2.0-FindIntentAppDRightContext) Should find intent 'aTestingIntent' belonging only to app intent-a with context 'testContextX'", async () => {
+    it("(FindIntentAppDRightContext) Should find intent 'aTestingIntent' belonging only to app intent-a with context 'testContextX'", async () => {
       const appIntent = await fdc3.findIntent(Intent.aTestingIntent, { type: ContextType.testContextX });
       validateAppIntent(
         appIntent,
@@ -40,7 +40,7 @@ export default async () =>
       );
     });
 
-    it('(2.0-FindIntentAppDWrongContext) Should throw NoAppsFound error when intent exists but context does not', async () => {
+    it('(FindIntentAppDWrongContext) Should throw NoAppsFound error when intent exists but context does not', async () => {
       try {
         await fdc3.findIntent(Intent.aTestingIntent, { type: ContextType.testContextY });
         assert.fail('No error was thrown', findIntentDocs);
@@ -49,7 +49,7 @@ export default async () =>
       }
     });
 
-    it("(2.0-FindIntentAppDMultiple1) Should find intent 'sharedTestingIntent2' belonging to multiple apps (intent-a & intent-b)", async () => {
+    it("(FindIntentAppDMultiple1) Should find intent 'sharedTestingIntent2' belonging to multiple apps (intent-a & intent-b)", async () => {
       const appIntent = await fdc3.findIntent(Intent.sharedTestingIntent2);
       validateAppIntent(
         appIntent,
@@ -59,7 +59,7 @@ export default async () =>
       );
     });
 
-    it("(2.0-IntentAppDMultiple2) Should find intent 'sharedTestingIntent2' belonging to multiple apps (intent-a & intent-b) filtered by specific context 'testContextY'", async () => {
+    it("(IntentAppDMultiple2) Should find intent 'sharedTestingIntent2' belonging to multiple apps (intent-a & intent-b) filtered by specific context 'testContextY'", async () => {
       const appIntent = await fdc3.findIntent(Intent.sharedTestingIntent2, { type: ContextType.testContextY });
       validateAppIntent(
         appIntent,
@@ -69,7 +69,7 @@ export default async () =>
       );
     });
 
-    it("(2.0-FindIntentAppDByResultSingle) Should find intent 'cTestingIntent' belonging only to app intent-c with context 'testContextX' and result type 'testContextZ'", async () => {
+    it("(FindIntentAppDByResultSingle) Should find intent 'cTestingIntent' belonging only to app intent-c with context 'testContextX' and result type 'testContextZ'", async () => {
       const appIntent = await fdc3.findIntent(
         Intent.cTestingIntent,
         { type: ContextType.testContextX },
@@ -83,7 +83,7 @@ export default async () =>
       );
     });
 
-    it("(2.0-FindIntentAppDByResultSingleNullContext) Should find intent 'cTestingIntent' belonging only to app intent-c with a null context and result type 'testContextZ'", async () => {
+    it("(FindIntentAppDByResultSingleNullContext) Should find intent 'cTestingIntent' belonging only to app intent-c with a null context and result type 'testContextZ'", async () => {
       const appIntent = await fdc3.findIntent(Intent.cTestingIntent, undefined, ContextType.testContextZ);
       validateAppIntent(
         appIntent,
@@ -93,7 +93,7 @@ export default async () =>
       );
     });
 
-    it("(2.0-FindIntentAppDByResultMultiple) Should find intent 'sharedTestingIntent1' belonging only to app intent-b with context 'testContextX' and result type 'testContextY'", async () => {
+    it("(FindIntentAppDByResultMultiple) Should find intent 'sharedTestingIntent1' belonging only to app intent-b with context 'testContextX' and result type 'testContextY'", async () => {
       const appIntent = await fdc3.findIntent(
         Intent.sharedTestingIntent1,
         { type: ContextType.testContextX },
@@ -107,7 +107,7 @@ export default async () =>
       );
     });
 
-    it("(2.0-FindIntentAppDByResultChannel1) Should find intent 'sharedTestingIntent2' belonging only to apps intent-e and itent-f with context 'testContextY' and result type 'channel", async () => {
+    it("(FindIntentAppDByResultChannel1) Should find intent 'sharedTestingIntent2' belonging only to apps intent-e and itent-f with context 'testContextY' and result type 'channel", async () => {
       const appIntent = await fdc3.findIntent(
         Intent.sharedTestingIntent2,
         { type: ContextType.testContextY },
@@ -121,7 +121,7 @@ export default async () =>
       );
     });
 
-    it("(2.0-FindIntentAppDByResultChannel2) Should find intent 'sharedTestingIntent2' belonging only to app intent-c with context 'testContextY' and result type 'channel<testContextZ>'", async () => {
+    it("(FindIntentAppDByResultChannel2) Should find intent 'sharedTestingIntent2' belonging only to app intent-c with context 'testContextY' and result type 'channel<testContextZ>'", async () => {
       const appIntent = await fdc3.findIntent(
         Intent.sharedTestingIntent2,
         { type: ContextType.testContextY },