+++ toolbox/fdc3-conformance/src/test/advanced/fdc3.getInfo.ts
@@ -1,12 +1,12 @@
 import { expect } from 'chai';
 import { handleFail, wrapPromise } from '../../utils';
-import { closeMockAppWindow } from '../fdc3-2_0-utils';
+import { closeMockAppWindow } from '../fdc3-conformance-utils';
 import { ImplementationMetadata, Listener } from '@finos/fdc3';
-import { MetadataValidator, MetadataContext, MetadataFdc3Api } from '../support/metadata-support-2.0';
-import { APIDocumentation2_0 } from '../support/apiDocuments-2.0';
-import { ControlContextType } from '../support/intent-support-2.0';
+import { MetadataValidator, MetadataContext, MetadataFdc3Api } from '../support/metadata-support';
+import { APIDocumentation } from '../support/apiDocuments';
+import { ControlContextType } from '../support/intent-support';
 
-const getInfoDocs = '\r\nDocumentation: ' + APIDocumentation2_0.getInfo + '\r\nCause';
+const getInfoDocs = '\r\nDocumentation: ' + APIDocumentation.getInfo + '\r\nCause';
 const validator = new MetadataValidator();
 const api = new MetadataFdc3Api();
 
@@ -22,7 +22,7 @@ export default async () =>
       }
     });
 
-    it('(2.0-GetInfo1) Returns a valid ImplementationMetadata object', async () => {
+    it('(GetInfo1) Returns a valid ImplementationMetadata object', async () => {
       try {
         const implMetadata = await api.getInfo();
         validator.validateImplementationMetadata(implMetadata);
@@ -31,7 +31,7 @@ export default async () =>
       }
     });
 
-    it('(2.0-GetInfo2) Returns a valid ImplementationMetadata object', async function () {
+    it('(GetInfo2) Returns a valid ImplementationMetadata object', async function () {
       this.timeout(10000);
 
       try {