+++ toolbox/fdc3-conformance/src/test/advanced/fdc3.open.ts
@@ -1,31 +1,29 @@
 import constants from '../../constants';
 import { openApp, OpenCommonConfig } from '../support/open-control';
-import { APIDocumentation2_0 } from '../support/apiDocuments-2.0';
-import { OpenControl2_0 } from '../support/open-support-2.0';
+import { APIDocumentation } from '../support/apiDocuments';
+import { OpenControlImpl } from '../support/open-support';
 import { DesktopAgent, getAgent } from '@finos/fdc3';
 import { assert, expect } from 'chai';
-import { ControlContextType } from '../support/intent-support-2.0';
+import { ControlContextType } from '../support/intent-support';
 
-const documentation = '\r\nDocumentation: ' + APIDocumentation2_0 + '\r\nCause:';
+const documentation = '\r\nDocumentation: ' + APIDocumentation + '\r\nCause:';
 
 const config: OpenCommonConfig = {
-  fdc3Version: '2.0',
-  prefix: '2.0-',
   target: 'AppIdentifier',
   targetMultiple: 'AppIdentifier',
 };
 
 export default async () => {
-  let control: OpenControl2_0;
+  let control: OpenControlImpl;
   let fdc3: DesktopAgent;
 
   beforeEach(async () => {
     fdc3 = await getAgent();
-    control = new OpenControl2_0(fdc3);
+    control = new OpenControlImpl(fdc3);
   });
 
   return describe('fdc3.open', () => {
-    const AOpensB3 = `(${config.prefix}AOpensB3) Can open app B from app A with no context and ${config.targetMultiple} as config.target`;
+    const AOpensB3 = `(AOpensB3) Can open app B from app A with no context and ${config.targetMultiple} as config.target`;
     it(AOpensB3, async () => {
       const targetApp = control.createTargetAppIdentifier(openApp.b.id);
       const result = control.contextReceiver('fdc3-conformance-opened');
@@ -34,7 +32,7 @@ export default async () => {
       await control.closeMockApp(AOpensB3);
     });
 
-    const AFailsToOpenB3 = `(${config.prefix}AFailsToOpenB3) Receive AppNotFound error when targeting non-existent app AppMetadata (${config.targetMultiple}) as config.target`;
+    const AFailsToOpenB3 = `(AFailsToOpenB3) Receive AppNotFound error when targeting non-existent app AppMetadata (${config.targetMultiple}) as config.target`;
     it(AFailsToOpenB3, async () => {
       try {
         const targetApp = control.createTargetAppIdentifier('ThisAppDoesNotExist');
@@ -45,7 +43,7 @@ export default async () => {
       }
     });
 
-    const AOpensBWithContext3 = `(${config.prefix}AOpensBWithContext3) Can open app B from app A with context and AppMetadata (${config.targetMultiple}) as target, app B adds generic listener`;
+    const AOpensBWithContext3 = `(AOpensBWithContext3) Can open app B from app A with context and AppMetadata (${config.targetMultiple}) as target, app B adds generic listener`;
     it(AOpensBWithContext3, async () => {
       const context = { type: 'fdc3.instrument', name: 'context' };
       const targetApp = control.createTargetAppIdentifier(openApp.b.id);
@@ -55,7 +53,7 @@ export default async () => {
       await control.closeMockApp(AOpensBWithContext3);
     });
 
-    const AOpensBWithSpecificContext = `(${config.prefix}AOpensBWithSpecificContext) Can open app B from app A with context and ${config.targetMultiple} as config.target and app B is expecting context`;
+    const AOpensBWithSpecificContext = `(AOpensBWithSpecificContext) Can open app B from app A with context and ${config.targetMultiple} as config.target and app B is expecting context`;
     it(AOpensBWithSpecificContext, async () => {
       const context = { type: 'fdc3.instrument', name: 'context' };
       const targetApp = control.createTargetAppIdentifier(openApp.b.id);
@@ -65,7 +63,7 @@ export default async () => {
       await control.closeMockApp(AOpensBWithSpecificContext);
     });
 
-    const AOpensBMultipleListen = `(${config.prefix}AOpensBMultipleListen) Can open app B from app A with context and ${config.targetMultiple} as config.target but app B has multiple listeners added before the correct one`;
+    const AOpensBMultipleListen = `(AOpensBMultipleListen) Can open app B from app A with context and ${config.targetMultiple} as config.target but app B has multiple listeners added before the correct one`;
     it(AOpensBMultipleListen, async () => {
       const context = { type: 'fdc3.instrument', name: 'context' };
       const targetApp = control.createTargetAppIdentifier(openApp.b.id);
@@ -76,7 +74,7 @@ export default async () => {
       await control.closeMockApp(AOpensBMultipleListen);
     });
 
-    const AOpensBWithWrongContext = `(${config.prefix}AOpensBWithWrongContext) Received App timeout when opening app B with fake context, app b listening for different context`;
+    const AOpensBWithWrongContext = `(AOpensBWithWrongContext) Received App timeout when opening app B with fake context, app b listening for different context`;
     it(AOpensBWithWrongContext, async () => {
       await control.addListenerAndFailIfReceived();
       const targetApp = control.createTargetAppIdentifier(openApp.b.id);